+++ src/models/Discount.php
@@ -391,7 +391,7 @@ class Discount extends Model
         $rules[] = [['orderConditionFormula'], 'string', 'length' => [1, 65000], 'skipOnEmpty' => true];
         $rules[] = [
             'orderConditionFormula', function($attribute, $params, $validator) {
-                if ($this->orderConditionFormula) {
+                if ($this->{$attribute}) {
                     $order = Order::find()->one();
                     if (!$order) {
                         $order = new Order();
@@ -399,7 +399,7 @@ class Discount extends Model
                     $orderDiscountConditionParams = [
                         'order' => $order->toArray([], ['lineItems.snapshot', 'shippingAddress', 'billingAddress'])
                     ];
-                    if (!Plugin::getInstance()->getFormulas()->validateConditionSyntax($this->orderConditionFormula, $orderDiscountConditionParams)) {
+                    if (!Plugin::getInstance()->getFormulas()->validateConditionSyntax($this->{$attribute}, $orderDiscountConditionParams)) {
                         $this->addError($attribute, Craft::t('commerce', 'Invalid order condition syntax.'));
                     }
                 }