+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+### Changed
+- Order condition forumlas now include serialized custom field values for use in formulas. ([#2066]https://github.com/craftcms/commerce/issues/2066))
+
 ### Fixed
 - Fixed a bug where shipping rules were enforcing a required “Order Condition Formula” field incorrectly. ([#2098](https://github.com/craftcms/commerce/issues/2098))
 - Fixed a bug where Base Discount and Per Item Discount fields would show negative values on the Edit Discount page. ([#2090](https://github.com/craftcms/commerce/issues/2090))